push

+++ src/api/ApiHistories.js
@@ -1,13 +1,10 @@
-import axios from "axios";
-
-export const api = axios.create({
-    baseURL: "http://localhost:8080/admin",
-});
+import api from "./apiConfig";
 
 export async function getAllHistory() {
     try {
         const token = localStorage.getItem("token");
-        const result = await api.get("/BookingHistory/Search", {
+
+        const result = await api.get("/admin/BookingHistory/Search", {
             headers: {
                 Authorization: `Bearer ${token}`,
             },
@@ -20,7 +17,13 @@ export async function getAllHistory() {
 
 export async function deleteHistory(historyId) {
     try {
-        const result = await api.delete(`/BookingHistory/Delete/${historyId}`);
+        const token = localStorage.getItem("token");
+
+        const result = await api.delete(`/admin/BookingHistory/Delete/${historyId}`, {
+            headers: {
+                Authorization: `Bearer ${token}`,
+            },
+        });
         return result.data.data;
     } catch (error) {
         throw new Error(`Có lỗi ${error.message}`);
@@ -29,7 +32,13 @@ export async function deleteHistory(historyId) {
 
 export async function getHistoryById(historyId) {
     try {
-        const result = await api.get(`/BookingHistory/Detail/${historyId}`);
+        const token = localStorage.getItem("token");
+
+        const result = await api.get(`/admin/BookingHistory/Detail/${historyId}`, {
+            headers: {
+                Authorization: `Bearer ${token}`,
+            },
+        });
         return result.data.data;
     } catch (error) {
         throw new Error(`Error fetching history ${error.message}`);
@@ -38,7 +47,7 @@ export async function getHistoryById(historyId) {
 
 export async function searchHistories(name) {
     try {
-        const result = await api.get(`/BookingHistory/Search?key=${name}`);
+        const result = await api.get(`/admin/BookingHistory/Search?key=${name}`);
         return result.data.data || [];
     } catch (error) {
         console.error(`Error searching histories: ${error.message}`);
@@ -48,7 +57,13 @@ export async function searchHistories(name) {
 
 export const confirmCompleted = async (listId) => {
     try {
-        const response = await api.get(`/BookingHistory/Accept/${listId}`);
+        const token = localStorage.getItem("token");
+
+        const response = await api.get(`/admin/BookingHistory/Accept/${listId}`, {
+            headers: {
+                Authorization: `Bearer ${token}`,
+            },
+        });
         if (response.status === 200 && response.data.data === true) {
             return true; // Xác nhận thành công
         } else {
